Add files via upload
This project involves the implementation of clustering with the K-Means algorithm using PySpark for the analysis of household electricity meter data. The data used in our
research is household electricity meter data obtained from measurements in a house in Sceaux, Paris, France, available on the UCI Machine Learning Repository website. This data contains information on electricity usage over 47 months from December 2006 to November 2010, with a total of 1,048,575 records. In this project, data preprocessing was carried out to select relevant variables, transform data, and clean data. The variables used focus on active energy submetering 1, sub-metering 2, and sub-metering 3, which represent electricity consumption in various areas: the kitchen, laundry room, and the usage of electric water heaters and air conditioning.